Ensure Supevisor/Plug-ins dev versions are higher than the last release
Currently the dev version might be considered older than the current
release (by conventional versioning, where a dev appendix is considered
the dev version of an upcoming release of the same version number):
2023.07.0 > 2023.07.dev2001

However, that is not our intention after we release a Supervisor or
Plug-in at a given month: Subsequent dev builds from the same month
should be considered newer.

Change the dev versioning to include the patch version, and increment
the patch version to make sure that the latest dev release is always
considered higher than the last actual release. With this change if
Supervisors current release is 2023.07.1, we would generate the
following dev version:
2023.07.0 < 2023.07.1.dev2001

If getting the last release fails, we fall back to the previous scheme,
but we do add teh patch version. So all dev version numbers now do
include the patch part by default (e.g. 2023.08.0.dev0100).