Add helpers/stale-issues as an action
There's currently a reusable workflow for marking issues as stale.
This adds a reusable action, which allows for running the task along
other helpers (e.g. most likely helpers/lock-issues) on the same runner
and with a bit more flexibility than what can be achieved with the
workflow. The inputs and steps stay the same as in the workflow.